# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusX API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/reports/{reportId}:
    patch:
      operationId: reviewAdminReport
      tags: [admin]
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewReportBody"
      responses:
        "200":
          description: Report status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportItem"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
        '500':
          description: Failed to generate upload URL.
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ReportItem:
      type: object
      required: [id, postId, serviceId, reason, status, createdAt, targetTitle]
      properties:
        id:
          type: integer
        postId:
          type: ["integer", "null"]
        serviceId:
          type: ["integer", "null"]
        reason:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        targetTitle:
          type: string
...
    ReviewReportBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [dismissed, reviewed]
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
        purpose:
          type: string
          enum: [post-image, service-image]
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...